[Fix #283] players의 초기 Ready socket Event 연결 완
