+++ components/mafia/RemoteParticipantTile.tsx
@@ -1,38 +1,22 @@
 import PlayerDieImages from "@/assets/images/player_die.svg";
 import useClickHandler from "@/hooks/useClickHandler";
-import useSocketOn from "@/hooks/useSocketOn";
 import { useDiedPlayer } from "@/store/game-store";
 import { useJobImageState } from "@/store/image-store";
-import { useActivePlayer, useIsRemoteOverlay, useOverLayActions, useReadyPlayers } from "@/store/overlay-store";
+import { useActivePlayer, useIsRemoteOverlay, useReadyPlayers } from "@/store/overlay-store";
 import S from "@/style/livekit/livekit.module.css";
-import { socket } from "@/utils/socket/socket";
 import { ParticipantTile, ParticipantTileProps, useEnsureTrackRef } from "@livekit/components-react";
 import Image from "next/image";
-import { useEffect } from "react";
 
 const RemoteParticipantTile = ({ trackRef }: ParticipantTileProps) => {
   const trackReference = useEnsureTrackRef(trackRef);
   const PlayerId = useActivePlayer();
+  const diedPlayers = useDiedPlayer();
   const imageState = useJobImageState();
   const isRemoteOverlay = useIsRemoteOverlay();
   const remoteReadyStates = useReadyPlayers();
   const { clickHandler } = useClickHandler();
-  const { setReadyPlayers } = useOverLayActions();
 
-  const diedPlayers = useDiedPlayer();
   const diedPlayer = diedPlayers.find((diedPlayer) => diedPlayer === trackReference.participant.identity);
-
-  //NOTE - players의 실시간 준비 상태 update
-  const sockets = {
-    setReady: (playerId: string, isReady: boolean) => {
-      setReadyPlayers(playerId, isReady);
-    }
-  };
-
-  useSocketOn(sockets);
-
-  //NOTE - remotePlayer의 Ready 초기 상태
-  socket.on("userInfo", () => {});
 
   if (!trackReference) {
     return null;
